Removing spurious \\s from the Timestamp type matcher.
* The regexp for timestamps wasn't detecting them correctly.  Should now be fixed.
* Made the regexps for timestamps and dates a little more forgiving.  single-digit months, days, hours, mins, etc... are now acceptable.
