+++ src/main/scala/mimir/algebra/Type.scala
@@ -82,8 +82,8 @@ object Type {
   val tests = Map[Type,Regex](
     TInt()   -> "^(\\+|-)?([0-9]+)$".r,
     TFloat() -> "^(\\+|-)?([0-9]*(\\.[0-9]+)?)$".r,
-    TDate()  -> "^[0-9]{4}\\-[0-9]{2}\\-[0-9]{2}$".r,
-    TTimestamp() -> "^[0-9]{4}\\-[0-9]{2}\\-[0-9]{2}\\ \\[0-9]{2}\\:[0-9]{2}\\:([0-9]{2}|[0-9]{2}\\.[0-9]*)".r,
+    TDate()  -> "^[0-9]{4}\\-[0-9]{1,2}\\-[0-9]{1,2}$".r,
+    TTimestamp() -> "^[0-9]{4}\\-[0-9]{1,2}\\-[0-9]{1,2}\\s+[0-9]{1,2}:[0-9]{1,2}:([0-9]{1,2}|[0-9]{0,2}\\.[0-9]*)$".r,
     TBool()  -> "^(?i:true|false)$".r
   )
   def matches(t: Type, v: String): Boolean =
