+++ test/data/DetectSeriesTest1.csv
@@ -1,0 +1,9 @@
+Tran_TS,Join_DT,Exp_DT,DOB
+2013-10-07 08:23:19.120,2013-10-07,2017-06-30,1990-11-07
+2014-06-15 08:23:19,2014-06-15,2017-06-30,1994-02-22
+2014-10-28 08:23:19.120,2014-10-28,2018-06-30,1993-12-16
+2015-08-17 08:23:19.120,2015-08-17,2020-06-30,1991-05-21
+2015-10-27 08:23:19.120,2015-10-27,2019-06-30,1995-08-16
+2015-11-06 08:23:19.120,2015-11-06,2019-06-30,1989-06-25
+2016-03-21 08:23:19.120,2016-03-21,2019-06-30,1992-02-09
+2017-04-10 08:23:19.120,2017-06-21,2021-06-30,1994-01-28
